revert(wiz): do not mark non-JDBC data sources as leaves in annotation tree
Refs #76075

Commit d3728a2b76 ("do not offer non-JDBC sources as expandable in the
annotation tree", PR #4581) was live-verified and reviewed twice, but was
found to break data labeling for genuinely-usable tabular sources (OData,
SAP, etc. -- these also extend TabularDataSource, same as truly-empty
sources like Cassandra) and was closed unmerged per that discussion.
However, that same commit was accidentally carried onto this branch's
mainline via an unrelated 146-commit integration merge and has been live
here since.

Non-JDBC/tabular data source support is under active development; the
user has directly confirmed marking these sources leaf (or hiding them)
is the wrong behavior and support is coming shortly. Separately, no
unhandled-exception risk exists for expanding one of these nodes today --
AbstractAssetEngine.getPhysicalEntries already has a pre-existing,
unrelated instanceof JDBCDataSource guard that returns an empty array
cleanly, so there was never anything to guard against here.

This reverts filterWizTree's isNonJdbcDataSource check and leaf(true)
branch, removes the now-dead isNonJdbcDataSource helper (single call
site), and deletes MetadataApiServiceNonJdbcTreeTest (all 7 tests were
specific to the reverted behavior). Adds a small regression test
(MetadataApiServiceFilterWizTreeTest) asserting a non-JDBC source is NOT
marked leaf, to guard against this exact reintroduction happening again
via a future merge.

No stylebi-wiz (portal) change needed -- confirmed nothing there reads
the leaf field for this tree; the closed companion PR (stylebi-wiz#1568)
never merged.
